improvements for AI and future documentation standards
extensive updates to the data source docs to include far better examples that AI can follow. Created documentation and coding standards primarily for AI to follow when updating documentation for other sections of API docs (not for guides or tutorials). The documentation and code standards haven't yet been tested with AI to confirm their effectiveness and this should be done as part of the code review